feat(healthz): refactor apiserver
